exports Cardz component to separate file and requires it in index.jsx
